validation file: fix archive validation overwrite
With 502442c7b ("dump: remove 2nd extract of archive") a new section
for validating the archive was introduced. This uses the ASCIIDocLog
class to append the new section. However, by default that class
creates a new file instead of appending.

Append to the validation file instead of re-creating it.

Rather than assume the ordering of check_full_pkgs() and
elbe_report(), change both functions to append and handle the
deletion of an old validation file separately.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>