elbeproject pbuilder: add gnupg to extrapackages installed into pbuilder
when using a pbuilder for a stretch target, we see this message:

--------------------------------------------------------------------
I: user script //tmp/hooks/D10elbe_apt_sources starting
E: gnupg, gnupg2 and gnupg1 do not seem to be installed, but one of them is required for this operation
--------------------------------------------------------------------

The result is that apt-key does not install the project repo key.
That makes builds which depend on packages build earlier fail.

Its not yet clear, if this also happens on elbe stable, and whether
<norecommend /> is involved (the reproducer has it, and is a stretch build)

Fix it by adding "gnupg" to  --extrapackages.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>

Signed-off-by: Manuel Traut <manut@linutronix.de>