dbsfed.xsd: change memory_restriction to unit prefixes
After moving the mem field from buildimage to initvm,
we are basically making a fresh start.
Drop the 128MB limit, and allow unit prefixes.

initvm/mem is only for experts. And it only
makes sense to specify it, when the default 1GiB is not enough.
So its unlikely that anybody will specify less than 128MB.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>