xmldefaults: add default for swap-size and fixup mem
when adding mem and swap-size, we need proper defaults, for xml
files not specifying these. The mem default was not evaluated prior
to this series, and we now specify memsize using unit prefixes
and size_to_int().

Change the per architecture mem sizes to 1GiB for all archs,
since this value is now only evaluated for the initvm and
pretty much arch independent.

swap-size defaults to "0"

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>