pbuilder: add --cpuset option to "elbe pbuilder build" and "elbe control"
qemu-user and java make problems with the current version of qemu
that is used.

-------------------------------------------------------------------------
qemu: uncaught target signal 6 (Aborted) - core dumped
/var/lib/dpkg/info/ca-certificates-java.postinst: line 43:

25111 Done                    find /etc/ssl/certs -name \*.pem
25113 Broken pipe             | while read filename; do
     alias=$(basename $filename .pem | tr A-Z a-z | tr -cs
              a-z0-9 _); alias=${alias%*_}; if [ -n "$FIXOLD" ]; then
         echo "-${alias}"; echo "-${alias}_pem";
     fi; echo "+${filename}";
 done
      25115 Aborted                 | java -Xmx64m -jar $JAR -storepass "$storepass"
 dpkg: error processing package ca-certificates-java (--configure):

 subprocess installed post-installation script returned error exit status 134
-------------------------------------------------------------------------

this seems to be an incarnation of this bug:
https://bugs.linaro.org/show_bug.cgi?id=3259

It seems to be fixed in a new qemu version. But upgrading to that version
is not so easy. Another possible mitigation is using only a single cpu.
In the past we used SMP=1 in the initvm Makefile for these kind of fixes.

Add an option that reduces the number of cpus using "taskset" per pbuild.
And put the necessary infrastructure in place to push the parameter down to
where it is evaluated.

This allows us to have a workaround in place for all classes of qemu-user
bugs with multithreading.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>