pkgutils: care about version numbers of packages
The current python-apt workaround code always downloads the last package
mentioned in the Packages file for a given package name. This is not
always the latest package.

This implements a minimalisic version comparision that is able to
download the latest elbe-bootstrap package and therefore fixes the
issue mentioned on github with the id #145.

However this will not work for all version strings that are allowed
in Debian.

This should be backported to elbe-2.x because as mentioned in #145
the issue is hard to decode.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>