schema: use all instead of sequence
'sequence' requires that each element appears multiple times or
not at all and cares on ordering of the elements. It also eases coding.
E.g. 'chg_archive' should insert the 'archive' on the correct position.
However the source-code always appends it to the end. That might lead to
validation errors, if there is another element before.

The ordering is not required by the elbe source code the parses the XML
for most cases and it's easier for humans if there is no required ordering
of the elements.

However, 'all' requires that each element appears only once or doesn't occure
at all. So it can't be used in all places that uses 'sequence' currently.
However in most places it is fine to use 'all' instead of 'sequence'.

This fixes issue #68 on github.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>