schema: set raw-preference to maxOccurs=1
currently the element can be inserted multiple times, the code handles
this correctly. However it avoids using the out 'all' tag instead of
'sequence' because 'all' just allows maxOccurs=1. Having a less strict
ordering in XML seems for me more useful than multiple occurance of the
rarely documented 'raw-preference' element. However this is a change of
the schema that makes it incompatible with the old one.

Since the raw-preference element supports multi-line data, there is no
benefit to supporting more than 1 occurrence.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>