elbexml: add check XML boolean helper
XML boolean attributes are specified as case sensitiv "true" or "1" for
a true and "false" or "0" for a false value. Other values are not
allowed and are rejected by validation. Non stated attributes defaults to
false.

elbexml (or better the underlaying lxml) refer None for non existing
attributes. Add a helper to check XML boolean values.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>