Use tar append to create archives
The new element <archivedir> which will be introduced in the next patch
may occur multiple times to generate an archive from a set of different
directories.

The existing chg_archive implementation uses the "-j" flag while calling
tar to compress the archive. Compressed archives can not be extended, but
the upcoming <archivedir> feature depends on it.

Generating the 'tar' archive needs to be split up into two parts:

a) generating the archive (may be called multiple times; implented in
   the collect function)

b) compressing the archive is not done with 'tar -j' but in python during
   enbasing the .tar archive.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>