Add archivedir support
The <archive> handling is unwieldy to track changes in the archive,
since it is not obvious, what changed in a base64 encoded bziped tar
archive.

Furthermore it has major drawbacks in variant handling:
If a variant needs changes in the archive the whole archive needs
to be exchanged. The changes between these archives are neither
obvious too.

The new element <archivedir> points to an local directory and add
the content into a newly created archive. <archivedir> can be
specified more then once. The content of the direcories is copied
in order of appearance. Existing files are overwritten by the later
ones. <archive> and <archivedir> are mutual exclusive.

<archivedir> is only allowed in XML files going into 'elbe
preprocess' and are converted into an <archive> tag by this elbe
subcommand.

Example snippet to use <archivedir>:
...
<archivedir>foo</archivedir>
<archivedir variant="production">bar</archivedir>
<archivedir keep-attributes="true">baz</archivedir>
<archivedir>file:///overlay</archivedir>
...

Keep in mind that URIs matching 'file://[^/].*' are rejected.
Handling these URIs is misleading at least and vary between
implementations.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>