docker: build without using local cache
Docker caches build steps. If e.g. mirrors are changed in a first step
not all necessary following steps are executed. The result is, that the
mirror change is in the sources.list, but the packages from the cached
run that was executed before is used.

Also due to the caching, updates on the debian mirror are ignored.

Signed-off-by: Manuel Traut <manut@linutronix.de>