pkgutils: update version string for elbe-debootstrap
With commit 4d69cfcbb864 ("pkgutils: care about version numbers of
packages"), version detection logic was implemented to always
choose the latest version. However, it relies on the package
version strings matching a certain pattern:

    0+deb0u0+jessie0

For the newer elbe-debootstrap packages, the pattern has changed.
This means that the downloader fallback code (and probably
installation from CDROM) is broken.

Adjust the pattern for the new version strings:

    0+deb0u0+deb8

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>