virtapt/lookup_uri: handle 'provides'
If going through the dependencies of a package 'provides' might
be used instead of a real package name.  If get_uri is used with
incl_deps=True lookup_uri can be called with a 'provides'
instead of a package name.

Extend lookup_uri with a loop that iterates over all packages
in the cache and check their 'provides'  if the package name
can't be found in the cache. If a 'provides' matches use the
according package to return its uri.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Volker Haspel <volker.haspel@linutronix.de>