pylint: daemons soap faults: fix Exception order in soap_faults()
InvalidState is derived from ProjectManagerError.
This must be caught first, or we would never Catch that.

reorder exceptions.

Note that we must do this for all argument count.
I now know, how to implement this more generically.

We could also delay this until all the cases are collapsed to one.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>