virtapt: remove elbe-repo.pub from repository
VirtApt.setup_gpg() inserts a public key into the keyring,
which is taken from inside the elbepack python module.
This is not necessary, because the key for the elbe repo is also
inserted into the virtapt keyring from the hosts debian keyrings,
and additionally via the key urls from the elbe xml file.

Remove the key and all references to it. Also remove it from
debian/*.install and setup.py.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>