pylint: fix dangerous arguments
To a Python novice, this may seem like a reasonable way to default to an
empty array for the extras param. What really happens is that this
"default" array gets created as a persistent object, and every
invocation of my_method that doesn't specify an extras param will be
using that same list object—any changes to it will persist and be
carried to every other invocation!

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>