soapclient/get_files: initialize retry variable
6b3fb9d01f170e708f357ea6c5f2b0e356ef0f3d removed the initiailization of
the retry variable, but not all usage. Because the commit
740690f56b0d79076face12ac365a18bde70af76 introducing the retry logic was
done for some reason, so we decided to keep the logic:
http://elbe-devel.2373613.n4.nabble.com/elbe-devel-PATCH-soapclient-download-file-fix-uninitialized-retry-tt1241.html#none
and add a verbose printout of the exception.

Also a comment is added, that the retry exception can be removed, if
those message didn't appear for a while.

This fixes github issue #171

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: John Ogness <jogness@linutronix.de>