init: dont tar up the ./initvm in devel mode
"./elbe initvm create --devel" creates the initvm in ./initvm, and
tars up the build directory. this produces:

tar: ./initvm/.elbe-in/elbe-devel.tar.bz2: file changed as we read it

this is not critical, but should be avoided.
Check whether the output directory lies in the devel directory, and
add a proper "--exclude" parameter.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>