debinstaller: implement download of vmlinuz and initrd.gz
the debinstaller module shall replace elbe-bootstrap.
it downloads debian installer linux kernel and initrd.gz from
a debian mirror. It does that in a secure manner, iE validating
Release.gpg and SHA256SUMS on the way.

filenames on the mirror:
/debian/dists/jessie/main/installer-amd64/current/images/netboot/debian-installer/amd64/linux
/debian/dists/jessie/main/installer-amd64/current/images/netboot/debian-installer/amd64/initrd.gz
/debian/dists/jessie/main/installer-amd64/current/images/cdrom/initrd.gz

Although this functionality is also provided by apt,
we implement it here in pure python, because outside
of the initvm, we can not rely on apt being available.

The initrd and vmlinuz are stored in the initvm in /var/cache/elbe/installer.
Also put them on bin-cdrom.iso, and reuse them, when an elbe build
is run from cdrom.

since elbe uses elbepack.debinstaller outside of the initvm now,
the non-virtap fallback code is not needed anymore.
Remove it and make pkgutils.py pylint clean.
Also rename the, now moved to debinstaller, NoKinitrdException
to NoPackageException.

Also remove code dealing with elbe bootstrap in packagelists and
xmldefaults.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>