fetch_initvm_pkgs: first version of fetch_initvm_pkgs
when the initvm is setup, it is necessary to download the sources,
used by the binary packages in the initvm. "elbe fetch_initvm_pkgs" is
supposed to build up a debian repo, that contains all the binaries and
sources used in the initvm.

Later on, this repo will be used to seed the initvm related packages and
sources in the generated cdroms.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>