repomanager: correct initvm related CdromSrcRepo attributes
Initvm uses components main and main/debian-installer. This is not
useful, because (contrary to binary Cdrom repos) initvm and target
share the source repo. This does not allow adding different package
versions for initvm and target, when arch and codename are identical.

Also, main/debian-installer does not contain sourcecode, and
additionally triggers a call to reprepro update during Repository
initialisation.

Rename component 'main' to 'initvm' for source repos. And drop
'main/debian-installer' completely.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>