init: init-elbe.sh.mako: call 'elbe buildrepo' as last thing
"elbe fetch_initvm_pkgs" is supposed to create package repositories which carry
binaries and sources used in the initvm. cdrom generation will use these
repos to a) speedup the build and b) allow for "offline" initvm.

call "elbe fetch_initvm_pkgs" as the last step during initvm install in
init-elbe.sh

cdrom has to be mounted, because its already unmounted, when init-elbe.sh
is executed. "in-target" can not be used to mount the cdrom, because
it cleans up mounts.

In case of a cdrom build specify "elbe fetch_initvm_pkgs" options:
"--cdrom-device /dev/sr0 --cdrom-mount-path /media/cdrom0"
This triggers a cdrom build.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>