init: add <max-cpus> field and make Makefile.mako obey the <mem> field
On machines with many cores, 1GiB of memory is not very much. Especially,
when g++ is used in a pbuilder with -jauto.

Add <max-cpus> field, that clamps the number of cpus for the initvm.
Make it default to 8. Also let Makefile.mako clamp the SMP default
to the max-cpus value.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>