install_elbe_version: implement "elbe control install_elbe_version"
when elbe on the hostmachine is upgraded, it usually can not talk
to the elbe version installed in the initvm.

Add a command that can install a specific elbe version in the initvm.
Inside the initvm apt will install a different version of elbe.
This is picked up by cherrypy. And the restart will happen,
after the asyncqueue is drained.

We have to specify the complete list of packages that need to be
installed at new versions, as the solver is not able to solve
the version locking.

--------------------------------------------------------------------------
The following packages have unmet dependencies:
 elbe-soap : Depends: elbe-daemon (= 2.9.11~bpo8+build251) but it is not going to be installed
 python-elbe-common : Depends: elbe-schema (= 2.9.11~bpo8+build251) but 2.9.12+deb8+build311 is to be installed
                      Depends: python-elbe-bin (= 2.9.11~bpo8+build251) but it is not going to be installed
E: Unable to correct problems, you have held broken packages.
--------------------------------------------------------------------------

Therefore the soap API requires that the list of packages to update is passed
from the host into the initvm.

To pass the String Array, the parameter type is declared
as String(max_occurs='unbounded'). According to
https://grokbase.com/t/python/soap/119vw5n35b/python-using-arrays-with-suds#20110927qbg2fy2fdoznyem5lc3xezkjsi
this works properly with suds.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>