fix /etc/hosts generation
fix github issue #185:

--8<--
The format of the /etc/hosts file for Debian Stretch is described here:
https://manpages.debian.org/stretch/manpages/hosts.5.en.html. According to this
document, the hostname
should be specified using 127.0.1.1 and not 127.0.01. There should be a tab
instead of a space between
127.0.1.1 and the hostname.

When giving an empty **<domain></domain>** section, the **/etc/hosts** file will
contain test.None. In
most cases we don't need/want a domain entry. There seems to be no way to tell
Elbe to skip generating a
domain entry.
--8<--

The statement about tabs as delimiters is wrong:
"Fields  of  the  entry are separated by any number of blanks and/or tab characters."

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>