finetuning: add a project-finetuning section
<project-finetuning> is a set of commands, that run after image generation.
It is designed  for extraction of partitions from disk-images, signing
artifacts, enable cryptsetups, ..

Add a method FinetuningAction.execute_prj() with an additional parameter
builddir. This method defaults to calling self.execute().
This allows using all existing FintuningActions in project-finetuning as well.
However, this must be allowed by the schema first. And care must be taken,
because not all FinetuningActions make sense to be called after Images have
been generated. It has no effect to modify target/ after the images have been
created.

Just allow <t2p_mv> for now. But remember that the files that are moved from
the target are still in the images!

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de> [reworked docs in schema and
commit message]