finetuning: add artifact and rm_artifact
Currently the set of files that are available for download from the initvm
is currently hardcoded. project-finetuning allows generation of new files.
It should be possible to mark this files for download from the initvm.

Therefore the <artifact> can be used.

It makes also sense to mark some files in normal finetuning as artifact. So
the usage of artifact is also allowed there.

rm_artifact is only allowed in project-finetuning and can be used to avoid
the download of files that elbe 'automaticly' considered to be downloaeded.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de> [rewrote commit message]