finetuning: implement <losetup> action
If a diskimage is mounted it becomes easy to e.g. extract partitions.

Therefore implement <losetup>. It is an action for <project-finetuning>.
It nests around other actions, which are ImageFinetuning actions.

These have an execute_img() method which adds
the name of the loop_device to the parameter list.

example:

<losetup img="sda.img">
   <extract_partition part="2">sda2.img</extract_partition>
</losetup>

The extract partition action is implemented in a followup patch.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>