init: fix xml to support jessie linux-base change
With commit 19a6a0cd6b59("init: bump initvm kernel version to
jessie-backports") linux-base from backports was added as a
dependency of the new kernel package. However, since then
linux-base was removed from backports and instead the security
package was updated from 3.5 to 4.5. This is quite unusual for
Debian and it broke elbe being able to create the initvm.

Remove the extra linux-base listing and add security as an apt
source so that the correct linux-base dependency can be found.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>