projectmanager: fix handling of utf-8 characters in logfile lines
ProjectManager.current_project_is_busy() returns str.
the result is returned as a String from ESoap.get_project_busy()
In reality spyne.model.primitive.String is Unicode, though.
When the return value can not be converted to Unicode, which is the
case, when a logline contains non ascii characters, the return value is
omitted from the xml result file.

This makes client.service.get_project_busy() return None in
WaitProjectBusyAction.execute() and this results in
the dubious "strange part: %d (skipped)" messages.

Fix this by calling logline.decode('utf-8','replace') and converting
the rest of the logline handling to unicode.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reported-by: Andreas Messerschmid <andreas.messerschmid@linutronix.de>
Tested-by: Andreas Messerschmid <andreas.messerschmid@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>