Replace gpgme with maintained gpg module
The pygpgme project is abandoned for years and not
available in the upcoming Debian buster.

Replace pygpgme with the official gpgme Python wrapper
from GnuPG that is available as Debian's python-gpg.
Replace the package references in README and the Debian
build. Use the low-level op_* functions to also target
the wrapper's predecessor pyme (Debian jessie).

With this change, elbe can be installed on Debian buster.
This closes issue #192.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Acked-by: Manuel Traut <manut@linutronix.de>