shellhelper: add env_add keyword arg to shellhelper.system()
shellhelper.system is a small wrapper around os.system() that makes
it behave like subprocess.check_call(). The difference is that it
throws shellhelper.CommandError and not subprocess.CalledProcessError.

add support for the env_add parameter, and now use subprocess.call()
instead of os.system() because that does not allow passing an environment.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>