initvm create: less strict XML check on project XML
On `elbe initvm create project.xml` with project.xml containing variants the
ElbeXML check on the XML file does not do preprocessing and fails.

However, later on the preprocessing is done anyway.
So just use an etree() call instead to do the XML check.
This closes issue #194.

Fix the user facing typo which appears in the bug report.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>