sdk: improve error handling and logging
None of the command executions in extract_pkg() are logged.
If a command fails fallbacks are implemented. But what is
actually executed can't be seen.

This makes the log parameter of extract_pkg non-optional
and use log.do() instead of system() to get command execution
logged.

It also fixups the callers to pass in a log object.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>