initvm: use stretch instead of jessie
build the initvm with Debian stretch instead of jessie.
jessie-backports was removed from the official debian mirrors.
The initvm used debootstrap from jessie-backports to be able to
build at least stretch targets.

To be still able to build targets with Debian/stretch the initvm
is updated to use stretch.

Debootstrap and pbuilder will be used from stretch-backports to
also support building buster targets.

The kernel is also used from stretch-backports.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>