initvmaction: fix race condition during initvm shutdown
Sometimes "elbe initvm stop" fails:

Traceback (most recent call last):
  File "/usr/bin/elbe", line 55, in <module>
    cmdmod.run_command(sys.argv[2:])
  File "/usr/lib/python2.7/dist-packages/elbepack/commands/initvm.py", line 91, in run_command
    action.execute(directory, opt, args[1:])
  File "/usr/lib/python2.7/dist-packages/elbepack/initvmaction.py", line 188, in execute
    raise e
libvirt.libvirtError: Requested operation is not valid: domain is not running

There are two states in libvirt. SHUTOFF that is used if the machine is
off and SHUTDOWN during the machine is going down.

The code sends a "shutdown" commmand to the machine in a loop.
After sending the shutdown command a machine is expected to be in the state
SHUTDOWN and than in the state SHUTOFF. Currently the code expects the
machine only to be in the SHUTOFF state, but not in the SHUTDOWN state.

Don't raise an error if the machine is in the SHUTDOWN state.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>