pkgutils: fix the LOCALMACHINE replace for in initvm code
pkgutils was used for elbe-bootstrap downloading code, which
runs outside of initvm. elbe-bootstrap has been removed, and different
code is used to download the debian installer nowadays.

The code in pkgutils is only used by the host sdk downloading code.
This code runs inside the initvm and needs to replace LOCALMACHINE
with 10.0.2.2

Fix it.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>