rfs: add key for local and cdrom repository
newer apt versions (like in stretch) verify the signature of copy
repositories.

Error Updating rpcaptcache: W:GPG error: copy:/repo stretch InRelease: The
following signatures couldn't be verified because the public key is not
available: NO_PUBKEY DCE4316BFFF45EAD, E:The repository 'copy:/repo stretch
InRelease' is not signed.

Therefore add the key of the localrepo and cdrom while entering the chroot and
remove it on exit.

The keys are only added/removed if the cdrom is available or the localrepo
contains packages.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Acked-by: Torben Hohn <torben.hohn@linutronix.de>