egpg: use gpg executable to export keys
The python-gpg bindings return the exported key in the keybox format.
python-pyme seemed to return it in gpg format.

The key is needed in gpg format for apt-key add.

I see no method to export the key in gpg format with python-gpg.
Therefore call the 'gpg' command to export they key.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>