rfs/debootstrap: gnupg is not bootstraped in buster
after debootstrapping a buster target without a debootstrap-variant gnupg is
missing in the target rfs.

however it is required by 'apt-key add' that is issued during entering the
chroot:

--8<--
running cmd +chroot /var/cache/elbe/7ec16c95-e147-425e-bb59-5e652bb1e734/chroot
apt-key add /repo/repo.pub+
------------------------------------------------------------------------------
E: gnupg, gnupg2 and gnupg1 do not seem to be installed, but one of them is
required for this operation
------------------------------------------------------------------------------

Command failed with errorcode 255
--8<--

Therefore add it to the debootstrap include.

This also removes the logging from building the debootstrap command, because
the command is printed during execution.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>