elbexml: rework repository validation a bit
every repository has been validated to caintain both binaries and
sources. This is not correct, when only a <binary> url is added.

Refactor the validation code, so that it checks whether r["binstr"]
exists and only adds the r["binstr"] when the correct validation
mode is used.

Since the validation mode is now implicit by adding the right values to
the repo dict, the parameter url_validation can now be removed from
the call to self.validate_repo(r)

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>