elbeproject: create the host_sysroot in a BuildEnv, like the target_sysroot
When more complex packages like qtcreator are to be installed into
the host sysroot, the extract_pkg() machinery fails to resolve the
dependencies properly. Error messages dont make any sense.

Use the new arch feature, to setup a BuildEnv for the host_sysroot.
This makes use of the normal debootstrap and rpc_aptcache machinery,
that is supposed to be used to generate rootfilesystems in elbe.

This is a preparation step for the change of the sysroots to subprojects,
with their own source.xml and source and bin cdroms.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>