init default-init.xml: remove backports kernel from pkg list
19a6a0cd6b59 ("init: bump initvm kernel version to jessie-backports")
switched the kernel to the jessie-backports kernel, to obtain
a 4.9 kernel to fix msgs like:

---------------------------------------------------------------
kvm_put_msrs: Assertion `ret == cpu->kvm_msr_buf->nmsrs'
---------------------------------------------------------------

stretch already has a 4.9 kernel, so this is not necessary anymore.

linux-image-amd64 from stretch-backports is currently subject to a
transition, and is broken. Since we dont need it, fall back to the
stretch kernel installed by the installer, and stay there.

examples/elbe-init-big-machine.xml is also fixed.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>