init init-elbe.sh.mako: name network interface host0 and use dhcp on it
the "persistent" naming of network interfaces is not persistent, when
different VM configurations come into play.

Somehow 33988ebe3d0b ("Enable initvm creation with newer qemu versions")
changed the network device name to enp4 during initvm installation in qemu.
When running with libvirt, it is enp3 again. That leaves the network
unconfigured.

Fix this once for all, by creating /etc/systemd/network/10-host.link
which matched the macaddr and renames the interface to host0.
Then use /etc/network/interfaces.d/host0 so that it gets configured using
dhcp.

This can potentially also fix a situation, where the initvm is run in
vmware or a different hypervisor.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>