makofiles toolchain-shar-extract.sh.mako: improve the relocation code
there is no distinction between target and native executables made.
All executables were patched with the native interpreter and x86_64
specific rpaths.

The relocation code also only handles executables and does not set the
rpath of shared objects.

Fix that by building lists of native_executables, native_elf_files (exes
and shared objects), and target_elf_files.
For elf files the rpath is setup, and executables also get the interpreter
setup.

The code also handles ld scripts like libpthread.so and libc6.so which
are relocated using "sed -i".

target_executable_files are not supported yet, because they need the
target ld-linux.so to exist. However, this case can not happen yet,
because the host sysroot can not be multiarch, yet.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>