Add definition/implementation of 'finetuning/file' XML tag
The tag 'file' can be used under the tag 'finetuning'.

The implementation will ensure that the path to the
destination exists by using 'mkdir_p'.

The user can provide the attribute 'append="true"' if she
wants to append to the destination instead of truncate it
(the default).

The user can also provide the attribute 'encoding'.  If
that's the case, the content of the XML tag is decode
according to the type of encoding before written to the
file.  As for now, the only encoding supported are plain
text (the default) and base64.

The user can provide the attribute 'owner', 'group' or
'mode' to the XML tag, using the 'chown', 'chgrp' and
'chmod' syntaxes.

NOTE!
=====

- The 'owner' and 'group' attributes might stop the
  operation if the user or group doesn't exist.  e.g. if the
  group doesn't exist and thus 'chgrp' failed, 'chmod' will
  never be done because of the exception throwed by Python

- The whole operation can fail if the destination is an
  existing directory

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>