Add save excursion in ChRootFilesystem context
Instead of manually copying/moving/deleting files around,
the '__enter__' method of 'ChRootFilesystem' should save the
excursion into 'self.excursion'.

The excursion can be hardcoded directly at the top of the
'__enter__' method.  Depending on the filesystem object,
other files can be add to the excursion as well dynamically.

To put a file into the excursion, 2 variables must be
provide and 1 optional.

* 'path' - file on the *host* to copy to the RFS.
* 'restore' - if to restore 'path' in the RFS on '__exit__'
* 'dst' (optional) - Copy 'path' to 'dst'

If 'dst' is not provided, then 'path' is copy to the same
place in the RFS.

Uppon exiting the 'ChrootFilesystem' context with '__exit__'
the excursion is restored according to the policies in it.
Finally, the excursion is deleted so it's reset in the next
context.

Examples
========

````````````````````````````````````````````````````````````
Excursion.add(self, "/usr/bin")
````````````````````````````````````````````````````````````

On '__enter__'

* Move "/rfs/usr/bin" to "/rfs/usr/bin.orig", even if it's a
  broken link

* Copy "/usr/bin" to "/rfs/usr/bin"

On '__exit__'

* Remove "/rfs/usr/bin"

* Move "/rfs/usr/bin.orig" to "/rfs/usr/bin"

````````````````````````````````````````````````````````````
ui = "/usr/share/elbe/qemu-elbe/elbe"
Excursion.add(self, ui, False, "/usr/bin")
````````````````````````````````````````````````````````````

On '__enter__'

* Move "/usr/share/elbe/qemu-elbe/elbe" to "/rfs/usr/bin/elbe"

On '__exit__'

* Remove "/rfs/usr/bin/elbe"

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>