Async logging for rpcaptcache
Before the RPC is started, its stdout and stderr are redirect to an
async logging thread that will ensure that _all_ outputs from the RPC
and the RPC children are properly catch.  See async_logging for more
details.

It's thus ok to use /print/ statements _only_ in the methods of
RPCAPTCache since everything is redirect to the logging thread.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>