Busy wait read from project log queue
Instead of reading the log file of a project, the wait_busy action
will now pull the project' message queue (see QHandler in "log.py").

If there's no message, wait_busy sleep for 0.1 second before trying
again.  Using the queue and sleeping should reduce the CPU usage of
the user drastically.

The special string "ELBE-FINISH", which shoudn't in theory appear as a
single line logged message, act as a sentinel to terminate the
wait_busy loop.  Of course if the SOAP server could allow to return
back a boolean and string that would be awesome .. but this hack work
for now.

* NOTE

  If multiple wait_busy instances try to pull from the *same* project'
  message queue, then they will fight for messages and can result in
  no message for one of the instance even though the project is
  producing outputs.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>